immersive-reader/app/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
azure-ai-projects==2.0.0
azure-identity==1.24.0
openai
httpx==0.28.1